fix(botones): corregir detección de flanco y bugs de firmware PIC16F877A
PROBLEMA PRINCIPAL

fix(botones): rb0_prev/rb1_prev/rb2_prev se sobreescribían antes de

evaluar flancos, por lo que las condiciones -pulsado nuevo- y -suelto

tras pulsado- nunca se cumplían. Se introduce rb0_was/rb1_was/rb2_was

para capturar el estado anterior antes de actualizar _prev. Afectaba:

pulsación corta de MODE (vistas, CONFIG, salida de MANUAL) y UP/DOWN

en MANUAL y CONFIG.

OTROS CAMBIOS

fix(adc): aplicar máscara & 0x03 sobre ADRESH en leer_ADC() para

descartar bits inválidos con ADFM=1 (justificado a la derecha).

fix(botones): cambiar tipo de advertencia_count de uint8_t a uint16_t

en config.h y main.c. El temporizador de -VALOR INVALIDO- comparaba

contra 2000 con un tipo de rango máximo 255 — nunca expiraba.

chore(build): excluir lcd.c del build (CMake y MPLAB X). El driver

LCD paralelo no se usa en tiempo de ejecución y comparte pines RB0-RB2

con los botones. El archivo permanece en disco sin modificaciones.

chore(vscode): añadir .vscode/c_cpp_properties.json con defines XC8

(__XC8, _XTAL_FREQ, __PIC16F877A__) y ruta al compilador xc8-cc.exe

v3.10 para reducir falsos positivos de IntelliSense.

chore(vscode): añadir --suppress-diag=main_returns_nonint a

clangd.arguments en settings.json. void main(void) es válido en XC8

para PIC16; el error era exclusivo del analizador clangd.

SIN CAMBIOS

- ticks_timer no se lee fuera de la ISR; no requiere protección atómica.

- Direcciones EEPROM ya eran consistentes con los defines de config.h.

- lcd.c y lcd.h permanecen en disco intactos.

+++ config.h
@@ -80,6 +80,6 @@ extern float temperatura_actual;
 
 // Advertencia CONFIG
 extern uint8_t advertencia_activa;
-extern uint8_t advertencia_count;
+extern uint16_t advertencia_count;
 
 #endif
